renamed 'stewi' from 'stewicombo_to_sector'
