+++ flowsa/flowbysector.py
@@ -47,7 +47,7 @@ from flowsa.EIA_MECS import mecs_energy_fba_cleanup, eia_mecs_energy_clean_alloc
     mecs_land_fba_cleanup, mecs_land_fba_cleanup_for_land_2012_fbs, mecs_land_clean_allocation_mapped_fba_w_sec
 from flowsa.EPA_NEI import clean_NEI_fba, clean_NEI_fba_no_pesticides
 from flowsa.StatCan_IWS_MI import convert_statcan_data_to_US_water_use, disaggregate_statcan_to_naics_6
-from flowsa.stewicombo_to_sector import stewicombo_to_sector, stewi_to_sector
+from flowsa.stewi import stewicombo_to_sector, stewi_to_sector
 from flowsa.USDA_CoA_Cropland import disaggregate_coa_cropland_to_6_digit_naics, coa_irrigated_cropland_fba_cleanup
 from flowsa.USDA_ERS_MLU import allocate_usda_ers_mlu_land_in_urban_areas, allocate_usda_ers_mlu_land_in_rural_transportation_areas
 from flowsa.USDA_IWMS import disaggregate_iwms_to_6_digit_naics
